[Gold V] Title: 물병, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Gold/1052. 물병/물병.cc
@@ -5,7 +5,7 @@ int main() {
     ios::sync_with_stdio(false);
     cin.tie(nullptr);
 
-    long long N;
+    int N;
     int K;
     cin >> N >> K;
 
@@ -15,10 +15,10 @@ int main() {
         return 0;
     }
 
-    long long added = 0;
+    int added = 0;
     // popcount(N) > K인 동안 가장 작은 1비트를 구매해서 합치기
     while (__builtin_popcountll(N) > K) {
-        long long lsb = N & -N;  // 최하위 1비트
+        int lsb = N & -N;  // 최하위 1비트
         added += lsb;            // 그만큼 물병 구매
         N += lsb;                // 구매 후 합치기 효과 반영
     }
